Remove CustomerTypeExtension, as the whole logic is only used in the Admin
