--- src/Sylius/Bundle/CoreBundle/Form/Extension/CustomerTypeExtension.php
@@ -1,41 +1,0 @@
-<?php
-
-/*
- * This file is part of the Sylius package.
- *
- * (c) Sylius Sp. z o.o.
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Sylius\Bundle\CoreBundle\Form\Extension;
-
-use Sylius\Bundle\CoreBundle\Form\EventSubscriber\AddUserFormSubscriber;
-use Sylius\Bundle\CoreBundle\Form\Type\User\ShopUserType;
-use Sylius\Bundle\CustomerBundle\Form\Type\CustomerType;
-use Symfony\Component\Form\AbstractTypeExtension;
-use Symfony\Component\Form\FormBuilderInterface;
-
-trigger_deprecation(
-    'sylius/core-bundle',
-    '1.14',
-    'The "%s" class is deprecated and will be removed in Sylius 2.0.',
-    CustomerTypeExtension::class,
-);
-
-/** @deprecated since Sylius 1.14 and will be removed in Sylius 2.0. */
-final class CustomerTypeExtension extends AbstractTypeExtension
-{
-    public function buildForm(FormBuilderInterface $builder, array $options): void
-    {
-        $builder->addEventSubscriber(new AddUserFormSubscriber(ShopUserType::class));
-    }
-
-    public static function getExtendedTypes(): iterable
-    {
-        return [CustomerType::class];
-    }
-}
