+++ src/Sylius/Bundle/CoreBundle/Resources/config/services/form.xml
@@ -96,9 +96,6 @@
         <service id="sylius.form.extension.type.taxon" class="Sylius\Bundle\CoreBundle\Form\Extension\TaxonTypeExtension">
             <tag name="form.type_extension" extended-type="Sylius\Bundle\TaxonomyBundle\Form\Type\TaxonType" priority="100" />
         </service>
-        <service id="sylius.form.extension.type.customer" class="Sylius\Bundle\CoreBundle\Form\Extension\CustomerTypeExtension">
-            <tag name="form.type_extension" extended-type="Sylius\Bundle\CustomerBundle\Form\Type\CustomerType" priority="100" />
-        </service>
         <service id="sylius.form.extension.type.promotion" class="Sylius\Bundle\CoreBundle\Form\Extension\PromotionTypeExtension">
             <tag name="form.type_extension" extended-type="Sylius\Bundle\PromotionBundle\Form\Type\PromotionType" priority="100" />
         </service>
